Pipeline done

+++ src/config/RawImageConfig.java
@@ -13,8 +13,8 @@ public class RawImageConfig {
     public boolean bigEndian;
 
     // --- Camps per a la compressió ---
-    public int qStep;            // El pas de quantització (Q)
-    public short[] frequencies;  // Taula de freqüències (Histograma)
+    public int qStep;
+    public int[] frequencies;  // ARRAY D'ENTERS
 
     public RawImageConfig(int width, int height, int bands, int bitsPerSample,
                           boolean signed, boolean bigEndian) {
@@ -26,20 +26,12 @@ public class RawImageConfig {
         this.bigEndian = bigEndian;
     }
 
-    /**
-     * Configura les dades necessàries per escriure la capçalera de compressió.
-     */
-    public void setCompressionHeaderData(int qStep, short[] frequencies) {
+    public void setCompressionHeaderData(int qStep, int[] frequencies) {
         this.qStep = qStep;
         this.frequencies = frequencies;
     }
 
-    /**
-     * Escriu el HEADER binari al flux de sortida (Encoder).
-     * @param dos El flux de dades de sortida.
-     */
     public void writeHeader(DataOutputStream dos) throws IOException {
-        // 1. Metadades Imatge
         dos.writeInt(width);
         dos.writeInt(height);
         dos.writeInt(bands);
@@ -47,27 +39,21 @@ public class RawImageConfig {
         dos.writeBoolean(signed);
         dos.writeBoolean(bigEndian);
 
-        // 2. Paràmetres de compressió
         dos.writeInt(qStep);
 
-        // 3. Taula de freqüències (Histograma)
+        // --- CORRECCIÓ CRÍTICA ---
+        // Fem servir writeInt (4 bytes) perquè les freqüències poden ser grans (fins a 2M)
         if (frequencies != null) {
-            dos.writeInt(frequencies.length); // Escrivim la mida de l'array
-            for (short freq : frequencies) {
-                dos.writeShort(freq);         // Escrivim cada freqüència
+            dos.writeInt(frequencies.length);
+            for (int freq : frequencies) {
+                dos.writeInt(freq); // writeInt, NO writeShort
             }
         } else {
             dos.writeInt(0);
         }
     }
 
-    /**
-     * Llegeix el HEADER binari d'un flux d'entrada i crea l'objecte de configuració (Decoder).
-     * @param dis El flux de dades d'entrada.
-     * @return L'objecte RawImageConfig carregat.
-     */
     public static RawImageConfig readHeader(DataInputStream dis) throws IOException {
-        // 1. Llegir metadades bàsiques
         int width = dis.readInt();
         int height = dis.readInt();
         int bands = dis.readInt();
@@ -77,15 +63,15 @@ public class RawImageConfig {
 
         RawImageConfig config = new RawImageConfig(width, height, bands, bitsPerSample, signed, bigEndian);
 
-        // 2. Llegir paràmetres de compressió
         config.qStep = dis.readInt();
 
-        // 3. Llegir taula de freqüències
+        // --- CORRECCIÓ CRÍTICA ---
+        // Fem servir readInt (4 bytes)
         int freqLength = dis.readInt();
         if (freqLength > 0) {
-            config.frequencies = new short[freqLength];
+            config.frequencies = new int[freqLength];
             for (int i = 0; i < freqLength; i++) {
-                config.frequencies[i] = dis.readShort();
+                config.frequencies[i] = dis.readInt(); // readInt, NO readShort
             }
         }
